feat: add configurations and migration. fix: models

+++ src/Domain/Tickets/Ticket.cs
@@ -3,21 +3,21 @@ using Domain.Sessions;
 
 namespace Domain.Tickets;
 
-public class Ticket(SessionId sessionId, int price, SeatId seatId)
+public class Ticket(SessionId sessionId, decimal price, SeatId seatId)
 {
     public SessionId SessionId { get; } = sessionId;
     public SeatId SeatId { get; } = seatId;
 
-    public int Price { get; private set; } = price;
+    public decimal Price { get; private set; } = price;
 
     public Session? Session { get; }
     public Seat? Seat { get; }
 
-    public void UpdateDetails(int price)
+    public void UpdateDetails(decimal price)
     {
         Price = price;
     }
 
-    public static Ticket New(SessionId sessionId, int price, SeatId seatId)
+    public static Ticket New(SessionId sessionId, decimal price, SeatId seatId)
         => new(sessionId, price, seatId);
 }
